feat(campaign): authoritative uint16 voxel path + namespaced block registry with frozen legacy table
- Voxel storage widened Uint8->Uint16 end to end (world store, generation,
  worker IPC, meshing, lighting); light/meta stay uint8
- blockRegistry is the single authority: frozen legacy 0-255 table locked by
  legacyIds.json snapshot test, dynamic allocation >=256, id 65535 reserved
  as the unknown-block placeholder (never silent air)
- Versioned codecs with legacy dual-read: acr body schema v2 (u16LE),
  export v3; electron CJS mirror kept byte-identical; enforcement of the
  unknown-id rule at renderer funnels (loadChunk, IDB readAll, import)
- WorldStorage.createWorld auto-attaches deterministic campaign graph +
  registry snapshot; saveWorldMeta preserves campaign fields
- Gate0 proving blocks (256-258) + atlas tiles 267-269 + placeholder tile 46
- Boots icon moved 238->125 (was sharing the echo-brick tile)

+++ src/systems/WorldManager.ts
@@ -1020,7 +1020,7 @@ export class WorldManager {
           // are cleared only AFTER the write succeeds; on failure they remain dirty
           // so the chunks are retried on the next pass (no silent data loss).
           const keys = Array.from(this.dirtyChunks);
-          const batch: Array<{ cx: number; cz: number; blocks: Uint8Array; light: Uint8Array; meta: Uint8Array }> = [];
+          const batch: Array<{ cx: number; cz: number; blocks: Uint16Array; light: Uint8Array; meta: Uint8Array }> = [];
           const savedKeys: Array<{ key: string; version: number }> = [];
           for (const key of keys) {
               const [cx, cz] = key.split(',').map(Number);
@@ -1564,7 +1564,7 @@ export class WorldManager {
     if (!chunk) return BlockType.AIR;
     return chunk[WorldCoords.index3D(lx, y, lz)];
   }
-  getChunkData(cx: number, cz: number, autoGenerate: boolean = true): Uint8Array | null {
+  getChunkData(cx: number, cz: number, autoGenerate: boolean = true): Uint16Array | null {
     const chunk = WorldStore.getChunkData(this.state, cx, cz);
     if (chunk) return chunk;
     if (autoGenerate && this.getStage(cx, cz) === ChunkStage.EMPTY) { this.queueGen(cx, cz, 0); }
